refactor: centralize api fetch utility
