+++ src/pages/ProviderSignup.tsx
@@ -9,8 +9,14 @@ import { Checkbox } from '@/components/ui/checkbox';
 import { CategorySelector } from '@/components/CategorySelector';
 import { Header } from '@/components/Header';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { ServiceCategory, VehicleType, requestOTP, verifyOTP, createProvider } from '@/lib/api';
-import { apiFetch } from '@/utils/api';
+import {
+  ServiceCategory,
+  VehicleType,
+  apiFetch,
+  requestOTP,
+  verifyOTP,
+  createProvider,
+} from '@/lib/api';
 import { Phone, Building, Radius, Clock, Truck, Bus } from 'lucide-react';
 import { useToast } from '@/hooks/use-toast';
